Backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.6.0
httpx>=0.27.0
openai>=1.55.0
pymilvus>=2.4.0
numpy>=1.26.0
pandas>=2.2.0
